+++ src/event.c
@@ -65,7 +65,7 @@ void cb_add_packet_cb(packet_cb cb) {
 	packet_cb_list = entry;
 };
 
-void cb_call_packet_cb(const int ptype, const uint8_t *packet, const int len, const char* ifname, const uint16_t vlanid) {
+void cb_call_packet_cb(const int ptype, const uint8_t *packet, const int len, const char* ifname, const int vlanid) {
 	for (struct packet_cb_list_entry* entry = packet_cb_list; entry; entry = entry->next) {
 		entry->cb(ptype, packet, len, ifname, vlanid);
 	}
