+++ src/fdb.h
@@ -27,7 +27,7 @@
 struct cache_fdb_entry
 {
 	char bridge[IF_NAMESIZE];
-	uint16_t vlanid;
+	int vlanid;
 	uint8_t mac[ETH_ALEN];
 	uint8_t enabled;
 	unsigned int portidx;
@@ -36,7 +36,7 @@ struct cache_fdb_entry
 
 typedef void (*update_fdb_cb)(struct cache_fdb_entry* entry, void* ctx);
 
-struct cache_fdb_entry* get_fdb_entry(const uint8_t* mac, const char* bridge, const uint16_t vlanid, const unsigned int portidx);
-struct cache_fdb_entry* add_fdb_entry(const uint8_t* mac, const char* ifname, const uint16_t vlanid, uint8_t enabled, unsigned int portidx);
+struct cache_fdb_entry* get_fdb_entry(const uint8_t* mac, const char* bridge, const int vlanid, const unsigned int portidx);
+struct cache_fdb_entry* add_fdb_entry(const uint8_t* mac, const char* ifname, const int vlanid, uint8_t enabled, unsigned int portidx);
 void update_fdb(update_fdb_cb, void* ctx);
 #endif
